* Account.cs: Statement,Delete,Deposit,withdrawal done.
* Program.cs:
* customer.cs:
* MainMenu.cs:
* LoginPage.cs:
* transHelper.cs:
* deleteAccount.cs:
* searchAccount.cs:
* transactAccount.cs:
* accountStatement.cs:
* BankingSystem.csproj:

+++ BankingSystem/customer.cs
@@ -9,6 +9,6 @@ namespace BankingSystem
         public string Email { get; set; }
         public long PhoneNumber { get; set; }
         public long AccountNumber { get; set; }
-        public int AccountBalance { get; set; }
+        public long AccountBalance { get; set; } 
     }
 }
